# docs/swagger.yaml
basePath: /api
definitions:
  models.ApiError:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
  models.ApiResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/models.ApiError'
      success:
        type: boolean
    type: object
...
  models.Role:
    properties:
      id:
        type: integer
      name:
        type: string
    required:
    - name
    type: object
  models.UserLoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.UserLoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/models.UserResponse'
    type: object
  models.UserRegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  models.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      role:
        $ref: '#/definitions/models.Role'
    type: object
host: localhost:8080
info:
  contact:
    email: bruno.malagoli@example.com
    name: Bsmart API Support
  description: RESTful API con soporte de WebSocket en tiempo real para gestión de
    productos y categorías. Sistema completo con autenticación JWT, control de acceso
    basado en roles, búsqueda full-text y seguimiento automático de historial de productos.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Bsmart Backend API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Autentica un usuario con email y contraseña, retorna un token JWT
        válido por 24 horas
      parameters:
      - description: Credenciales de login
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login exitoso, token JWT generado
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.UserLoginResponse'
              type: object
        "400":
          description: Datos de entrada inválidos
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "401":
          description: Credenciales inválidas
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "500":
          description: Error interno del servidor
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
      summary: Iniciar sesión
      tags:
      - autenticación
  /auth/register:
    post:
      consumes:
      - application/json
      description: Crea una nueva cuenta de usuario con rol "client" por defecto.
        El email debe ser único en el sistema.
      parameters:
      - description: Datos de registro
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UserRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Usuario creado exitosamente
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.UserLoginResponse'
              type: object
        "400":
          description: Datos de entrada inválidos
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "409":
          description: Email ya registrado
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "500":
          description: Error interno del servidor
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
      summary: Registrar nuevo usuario
      tags:
      - autenticación
...
securityDefinitions:
  BearerAuth:
    description: 'Autenticación JWT. Formato: "Bearer {token}". Obtén un token desde
      /api/auth/login o /api/auth/register'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"